Refactor tasks to create credentials rather than use pscale connect
By using credentials, we are able to make the rake tasks more reliable (some envs would fail using pscale connect).

We are also able to remove the need for people to have to configure `database.yml` specifically to work with this gem.

By setting `DATABASE_URL`, we override what is set in `database.yml`.