Fix sslca param + disable schema dumps for psdb
- sslca doesnt need an underscore

For the schema dump. I've found that when rails generates schema.rb or structure.sql, it runs an incredible amount of queries to construct dump.

Normally, with local mysql this is fine because there is nearly no latency per query. When running this against a cloud database, it adds up very quickly.

One of our example databases takes a full minute to dump (my laptop is 30ms from the database).

By adding `schema_dump=`, this sets the Rails config for schema_dump to nil. Telling Rails not to run the dump.

`schema_dump=false` does not work because this sets it to a string, which rails interprets at the filename where we want the dump stored.