better search for .pscale.yml file
In the case of monorepos, the .pscale.yml file may not be in the
current working directory as the pscale CLI will write it in the
Git root. This modifies the search logic to find the git root in
the event that the .pscale.yml file is not found in the CWD.

We are checking the use cases that the GIT_DIR is set or the user
is in a Git worktree as well.