Allow `create_table` to take a keyspace option for multi-keyspace databases
For databases with multiple keyspaces, all DDL is automatically routed to the correct keyspace with the exception of `CREATE TABLE`. For this, we need to tell vitess where we want the table to g.

This PR enables `create_table` to take a `keyspace` argument. Which is then prepended to the table name when running the migration.

This will only run when ENABLE_PSDB is set. This allows `create_table` to both work against PlanetScale/Vitess as well as local MySQL for development.