Add ability to set a default keyspace via an ENV var
For apps where most tables will always be added to a specific keyspace,
this makes it possible to set this keyspace via an ENV var PLANETSCALE_DEFAULT_KEYSPACE. Makes the dev workflow a bit easier and the keyspace only needs to be overridden when a table is being added
to a keyspace different from the default.

Also, AI'd up some basic test coverage.